Support Xcode 6. (breaks pre-Xcode 6)
Bunch of work in this change.
The main instruments-without-delay hook changed dramatically as Apple now has the Simulator be started by launchd. ScriptAgent is also started by DTMobileIS, which is in turn started by launchd_sim.
I removed our dependency on the clang-as-ios-dylib stuff as Xcode 6 now supports dynamic libraries for iOS.
Moved the Swizzle code to Common and added a tiny comment about it.